+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.0] - 2020-09-15
+
 ### Added
 
 - Introducing Kubernetes events
@@ -150,7 +152,8 @@ be edited by users.
 
 
 
-[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v3.1.1...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/giantswarm/cluster-operator/compare/v3.1.1...v3.2.0
 [3.1.1]: https://github.com/giantswarm/cluster-operator/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/giantswarm/cluster-operator/compare/v3.0.1...v3.1.0
 [3.0.1]: https://github.com/giantswarm/cluster-operator/compare/v3.0.0...v3.0.1